+++ packages/fluentui/react-northstar/src/components/Avatar/Avatar.tsx
@@ -7,17 +7,20 @@ import {
   useFluentContext,
   useStyles,
   useTelemetry,
+  mergeVariablesOverrides,
 } from '@fluentui/react-bindings';
 import * as customPropTypes from '@fluentui/react-proptypes';
 import * as PropTypes from 'prop-types';
 import * as React from 'react';
 
 import { Box, BoxProps } from '../Box/Box';
-import { Image, ImageProps } from '../Image/Image';
 import { Label, LabelProps } from '../Label/Label';
-import { Status, StatusProps } from '../Status/Status';
+
 import { ShorthandValue, FluentComponentStaticProps } from '../../types';
-import { createShorthandFactory, UIComponentProps, commonPropTypes, SizeValue } from '../../utils';
+import { createShorthandFactory, UIComponentProps, commonPropTypes, SizeValue, createShorthand } from '../../utils';
+import { AvatarStatusProps, AvatarStatus } from './AvatarStatus';
+import { AvatarImage, AvatarImageProps } from './AvatarImage';
+import { AvatarStatusIcon } from './AvatarStatusIcon';
 
 export interface AvatarProps extends UIComponentProps {
   /**
@@ -29,7 +32,7 @@ export interface AvatarProps extends UIComponentProps {
   icon?: ShorthandValue<BoxProps>;
 
   /** Shorthand for the image. */
-  image?: ShorthandValue<ImageProps>;
+  image?: ShorthandValue<AvatarImageProps>;
 
   /** Shorthand for the label. */
   label?: ShorthandValue<LabelProps>;
@@ -44,7 +47,7 @@ export interface AvatarProps extends UIComponentProps {
   size?: SizeValue;
 
   /** Shorthand for the status of the user. */
-  status?: ShorthandValue<StatusProps>;
+  status?: ShorthandValue<AvatarStatusProps>;
 
   /** Custom method for generating the initials from the name property, which is shown if no image is provided. */
   getInitials?: (name: string) => string;
@@ -56,7 +59,12 @@ export const avatarClassName = 'ui-avatar';
 /**
  * An Avatar is a graphical representation of a user.
  */
-export const Avatar: ComponentWithAs<'div', AvatarProps> & FluentComponentStaticProps<AvatarProps> = props => {
+export const Avatar: ComponentWithAs<'div', AvatarProps> &
+  FluentComponentStaticProps<AvatarProps> & {
+    Status: typeof AvatarStatus;
+    StatusIcon: typeof AvatarStatusIcon;
+    Image: typeof AvatarImage;
+  } = props => {
   const context = useFluentContext();
   const { setStart, setEnd } = useTelemetry(Avatar.displayName, context.telemetry);
   setStart();
@@ -95,12 +103,13 @@ export const Avatar: ComponentWithAs<'div', AvatarProps> & FluentComponentStatic
   const ElementType = getElementType(props);
   const unhandledProps = useUnhandledProps(Avatar.handledProps, props);
 
-  const imageElement = Image.create(image, {
+  const imageElement = createShorthand(AvatarImage, image, {
     defaultProps: () =>
       getA11Props('image', {
         fluid: true,
         avatar: !square,
         title: name,
+        // remove in upcoming breaking change
         styles: resolvedStyles.image,
       }),
   });
@@ -129,12 +138,16 @@ export const Avatar: ComponentWithAs<'div', AvatarProps> & FluentComponentStatic
     <ElementType {...getA11Props('root', { className: classes.root, ...unhandledProps })}>
       {hasGlyph && (imageElement || iconElement)}
       {!hasGlyph && labelElement}
-      {Status.create(status, {
+      {createShorthand(AvatarStatus, status, {
         defaultProps: () =>
           getA11Props('status', {
             size,
+            // remove in upcoming breaking change
             styles: resolvedStyles.status,
           }),
+        overrideProps: (predefinedProps: AvatarStatusProps) => ({
+          variables: mergeVariablesOverrides(variables, predefinedProps.variables),
+        }),
       })}
     </ElementType>
   );
@@ -186,6 +199,11 @@ Avatar.propTypes = {
   status: customPropTypes.itemShorthand,
   getInitials: PropTypes.func,
 };
+
+Avatar.Status = AvatarStatus;
+Avatar.StatusIcon = AvatarStatusIcon;
+Avatar.Image = AvatarImage;
+
 Avatar.handledProps = Object.keys(Avatar.propTypes) as any;
 
 Avatar.create = createShorthandFactory({ Component: Avatar, mappedProp: 'name' });