+++ packages/react-button/src/components/MenuButton/MenuButton.tsx
@@ -4,13 +4,14 @@ import { useInlineTokens } from '@fluentui/react-theme-provider';
 import { useMenuButton } from './useMenuButton';
 import { MenuButtonProps } from './MenuButton.types';
 import { useMenuButtonClasses } from './useMenuButtonClasses';
+import { renderMenuButton } from './renderMenuButton';
 
 /**
  * Define a styled MenuButton, using the `useMenuButton` hook.
  * {@docCategory Button}
  */
 export const MenuButton = React.forwardRef<HTMLElement, MenuButtonProps>((props, ref) => {
-  const { state, render } = useMenuButton(props, ref, {
+  const state = useMenuButton(props, ref, {
     menuIcon: { as: ChevronDownIcon },
   });
 
@@ -25,7 +26,7 @@ export const MenuButton = React.forwardRef<HTMLElement, MenuButtonProps>((props,
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   useInlineTokens(state as any, '--button');
 
-  return render(state);
+  return renderMenuButton(state);
 });
 
 MenuButton.displayName = 'MenuButton';