+++ packages/react-button/src/components/ToggleButton/useToggleButton.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { useButton } from '../Button/useButton';
-import { ToggleButtonProps } from './ToggleButton.types';
+import { ToggleButtonProps, ToggleButtonState } from './ToggleButton.types';
 import { useChecked } from './useChecked';
 
 export const useToggleButton = (
@@ -8,9 +8,9 @@ export const useToggleButton = (
   ref: React.Ref<HTMLElement>,
   defaultProps?: ToggleButtonProps,
 ) => {
-  const buttonDefinition = useButton(props, ref, defaultProps);
+  const state = useButton(props, ref, defaultProps);
 
-  useChecked(buttonDefinition.state as ToggleButtonProps);
+  useChecked(state as ToggleButtonProps);
 
-  return buttonDefinition;
+  return state as ToggleButtonState;
 };