+++ packages/react-cards/src/next/Card.types.ts
@@ -1,10 +1,13 @@
 import * as React from 'react';
 import { BaseSlots, SlotProps } from '@fluentui/react-compose';
 import { ComponentProps } from '@fluentui/react-compose/lib/next/index';
-import { ColorTokenSet, SizeValue } from '@fluentui/theme';
+import { ColorTokens, SizeValue } from '@fluentui/theme';
 
 /* eslint-disable @typescript-eslint/naming-convention */
 
+/**
+ * {@docCategory Card}
+ */
 export type CardProps = ComponentProps &
   React.HTMLAttributes<HTMLDivElement> & {
     /** A card can be compact, without any padding inside. */
@@ -20,7 +23,7 @@ export type CardProps = ComponentProps &
     size?: SizeValue;
 
     /** A card can take up the width and height of its container. */
-    fluid?: boolean;
+    block?: boolean;
 
     /** A card can show that it cannot be interacted with. */
     disabled?: boolean;
@@ -44,12 +47,70 @@ export type CardProps = ComponentProps &
     selected?: boolean;
   };
 
+/**
+ * {@docCategory Card}
+ */
 export interface CardState extends CardProps {
   cardRef: React.RefObject<HTMLDivElement>;
 }
 
+/**
+ * {@docCategory Card}
+ */
 export interface CardSlots extends BaseSlots {}
 
+/**
+ * {@docCategory Card}
+ */
 export type CardSlotProps = SlotProps<CardSlots, CardProps, React.HTMLAttributes<HTMLDivElement>>;
 
-export type CardTokens = ColorTokenSet & {};
+type SizeRelatedTokens = {
+  borderRadius?: string;
+  height?: string;
+  margin?: string;
+  padding?: string;
+  width?: string;
+};
+
+type StateChangeRelatedTokens = {
+  borderWidth?: string;
+  boxShadow?: string;
+  cursor?: string;
+};
+
+/**
+ * {@docCategory Card}
+ */
+export type CardTokens = ColorTokens &
+  SizeRelatedTokens &
+  StateChangeRelatedTokens & {
+    borderStyle?: string;
+    minHeight?: string;
+    minWidth?: string;
+
+    /* sizing */
+    size?: {
+      smallest?: SizeRelatedTokens;
+      smaller?: SizeRelatedTokens;
+      small?: SizeRelatedTokens;
+      medium?: SizeRelatedTokens;
+      large?: SizeRelatedTokens;
+      larger?: SizeRelatedTokens;
+      largest?: SizeRelatedTokens;
+    };
+
+    disabled?: StateChangeRelatedTokens;
+    hovered?: StateChangeRelatedTokens;
+    pressed?: StateChangeRelatedTokens;
+    selected?: StateChangeRelatedTokens;
+  };
+
+/**
+ * {@docCategory Card}
+ */
+export type CardVariants<TTokens = CardTokens> = {
+  root?: TTokens;
+  onClick?: TTokens;
+  compact?: TTokens;
+  block?: TTokens;
+};